+++ entity/src/main/java/boozelogger/entity/Vessel.java
@@ -14,7 +14,7 @@ import java.util.Date;
 @Table(name="vessel")
 public class Vessel {
 
-    private Long id;
+    private Integer id;
     private String name;
     private String description;
     private Double volume;
@@ -25,7 +25,7 @@ public class Vessel {
         this(null, null, null, null, null, null);
     }
 
-    public Vessel(Long id, String name, String description, Double volume, UnitOfMeasurement unit, Date createdAt) {
+    public Vessel(Integer id, String name, String description, Double volume, UnitOfMeasurement unit, Date createdAt) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -37,11 +37,11 @@ public class Vessel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonProperty
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
